cleanup the mmap_map

+++ mmap_map/tests/char_test.cpp
@@ -125,7 +125,7 @@ void test5(size_t n) {
 
     const auto str = it.get_name();
 
-    assert(std::atoi(std::string(str).c_str()) == conta);
+    assert(std::atoi(std::string(str).c_str()) == (int)conta);
 
     const auto &d = it.get_field();
     assert(d.potato == (0xdead0000 | (conta & 0xFFFF)));
